requirements.txt
tensorflow
pandas
numpy
dvc
notebook
Flask
notebook
Flask-Cors
dvc
matplotlib
seaborn
python-box==6.0.2
ensure==1.0.2
tqdm
PYaml
joblib
types-PyYAML
ipykernel
jupyter
scipy
# -e .